Refactor: Move file operations to FileRepository and update FileController to use IFileRepository. Removed file logic from CouchRepository.
