+++ PharmaCoreApi/Controllers/FileController.cs
@@ -13,11 +13,11 @@ namespace PharmaCoreApi.Controllers
     public class FileController : ControllerBase
     {
         private readonly ILogger<FileController> _logger;
-        private readonly ICouchRepository _couchRepository;
-        public FileController(ILogger<FileController> logger, ICouchRepository couchRepository)
+        private readonly IFileRepository _fileRepository;
+        public FileController(ILogger<FileController> logger, IFileRepository fileRepository)
         {
             _logger = logger;
-            _couchRepository = couchRepository;
+            _fileRepository = fileRepository;
         }
 
         [HttpPost]
@@ -28,7 +28,7 @@ namespace PharmaCoreApi.Controllers
             _logger.LogTrace("writing to text file data {0} {1}", drugDetails.DrugName, drugDetails.DrugExpiredOn, drugDetails.LotNo);
 
             string filePath = @"c:\FileData\Filedata.txt";
-            var result = await _couchRepository.WriteTextAsync(filePath, drugDetails);
+            var result = await _fileRepository.WriteTextAsync(filePath, drugDetails);
             if (result.Length > 0)
             {
                 _logger.LogInformation("data written to file sucessfully");
