+++ PharmaCoreApi/DBContext/CouchRepository.cs
@@ -245,28 +245,6 @@ namespace PharmaCoreApi.Models
             return response;
         }
 
-        public async Task<string> WriteTextAsync(string filePath, DrugDetails drugDetails)
-        {
-            bool isAddedOrAppended = false;
-            string text = drugDetails.DrugName + " " + drugDetails.DrugExpiredOn  + drugDetails.LotNo + Environment.NewLine;
-
-            string SAMPLE_KEY = "gCjK+DZ/GCYbKIGiAt1qCA==";
-            string SAMPLE_IV = "47l5QsSe1POo31adQ/u7nQ==";
-
-            var key = Encoding.UTF8.GetBytes(SAMPLE_KEY);
-            var iv = Encoding.UTF8.GetBytes(SAMPLE_IV);
-
-            Array.Resize(ref key, 128 / 8);
-            Array.Resize(ref iv, 128 / 8);
-
-            if (File.Exists(filePath) && File.ReadLines(filePath).Any(line => line.Length > 0))
-                isAddedOrAppended = await StreamHelper.AppendStringToFile(filePath, text, key, iv);
-            else
-                isAddedOrAppended = await StreamHelper.WriteStringToFile(filePath, text, key, iv);
-
-            return isAddedOrAppended ? "success" : "failure";
-        }
-
         //private HttpClient DbHttpClient()
         //{
         //    var httpClient = this._clientFactory.CreateClient();
