+++ PharmaCoreApi/DBContext/ICouchRepository.cs
@@ -15,7 +15,6 @@ namespace PharmaCoreApi.Models
         Task<HttpClientResponse> GetAllDocumentsAsync();
         Task<HttpClientResponse> GetViewAsync(QueryView key);
         Task<HttpClientResponse> DeleteDocumentAsync(string id, string rev);
-        Task<string> WriteTextAsync(string filePath, DrugDetails drugDetails);
 
 
     }
