+++ PharmaCoreApi/DBContext/IFileRepository.cs
@@ -1,0 +1,10 @@
+using System.Threading.Tasks;
+using PharmaCoreApi.Models;
+
+namespace PharmaCoreApi.Models
+{
+    public interface IFileRepository
+    {
+        Task<string> WriteTextAsync(string filePath, DrugDetails drugDetails);
+    }
+}
